refactor(ffmpeg): move and simplify buildCompressionCommand
Move buildCompressionCommand from recorder to ffmpeg service and simplify
its signature. The function is now co-located with its usage and has a
cleaner interface that requires explicit output path.
